+++ lib/src/helpers/helpers.exports.dart
@@ -1,4 +1,2 @@
 export './entry_helper.dart';
-export './extensions.dart';
 export './prints_helper.dart';
-export './usage_printer.dart';