+++ lib/src/public.exports.dart
@@ -1,3 +1,2 @@
 export './commands_tree.dart';
-export './helpers/usage_printer.dart';
 export './helpers/entry_helper.dart';